+++ handlers/zips.go
@@ -1,6 +1,8 @@
 package handlers
 
 import (
+	"net/http"
+
 	"github.com/gin-gonic/gin"
 	"github.com/spo-iitk/ras-cdn/db"
 	"github.com/spo-iitk/ras-cdn/utils"
@@ -15,11 +17,11 @@ type ZipRequest struct {
 func ZipFilesHandler(ctx *gin.Context) {
 	var req ZipRequest
 	if err := ctx.BindJSON(&req); err != nil {
-		ctx.AbortWithStatusJSON(400, gin.H{"error": err.Error()})
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
 		return
 	}
 	if f := db.CheckFilesZipExists(req.Files); f != "" {
-		ctx.JSON(200, gin.H{
+		ctx.JSON(http.StatusOK, gin.H{
 			"message":  "zipped",
 			"filename": f,
 		})
@@ -28,7 +30,7 @@ func ZipFilesHandler(ctx *gin.Context) {
 
 	uuid, err := utils.GenerateUUID()
 	if err != nil {
-		ctx.AbortWithStatusJSON(400, gin.H{"error": err.Error()})
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
 		return
 	}
 
@@ -38,7 +40,7 @@ func ZipFilesHandler(ctx *gin.Context) {
 		utils.ZipFiles(req.Files, x)
 	}()
 
-	ctx.JSON(200, gin.H{
+	ctx.JSON(http.StatusOK, gin.H{
 		"message":  "zipped",
 		"filename": x,
 	})
@@ -53,29 +55,41 @@ func DownloadZipHandler(ctx *gin.Context) {
 
 type DeleteZipRequest struct {
 	Filename string `json:"filename"`
+	Secret   string `json:"secret"`
 }
 
 func DeleteOneZipHandler(ctx *gin.Context) {
 	var req DeleteZipRequest
 	if err := ctx.BindJSON(&req); err != nil {
-		ctx.AbortWithStatusJSON(400, gin.H{"error": err.Error()})
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
 		return
 	}
 
 	filename := req.Filename
 
+	if req.Secret != secret {
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
+	}
+
 	db.DeleteZipRow(filename)
 
 	ok := utils.DeleteFile(zip_folder + "/" + filename)
 	if !ok {
-		ctx.AbortWithStatusJSON(400, gin.H{"error": "could not delete file"})
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "could not delete file"})
 		return
 	}
 }
 
 func DeleteZipsHandler(ctx *gin.Context) {
+	delete_secret := ctx.GetHeader("secret")
+	if delete_secret != secret {
+		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
+	}
+
 	utils.DeleteZips()
-	ctx.JSON(200, gin.H{
+	ctx.JSON(http.StatusOK, gin.H{
 		"message": "zips cleared",
 	})
 }
